% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca_split.R
\name{dim.split_input}
\alias{dim.split_input}
\title{Return Dimensions of each Chunk}
\usage{
\method{dim}{split_input}(x)
}
\arguments{
\item{x}{a \code{\link[=split_input]{split_input()}} object}
}
\value{
a list of dimensions of each chunk
}
\description{
Wrapper for \code{lapply(x, dim)}
}
