% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca_split.R
\name{print.split_input}
\alias{print.split_input}
\title{Print a \code{\link[=split_input]{split_input()}} Object}
\usage{
\method{print}{split_input}(x, n = 6L, k = 6L)
}
\arguments{
\item{x}{a \code{\link[=split_input]{split_input()}} object}

\item{n}{number of rows to print}

\item{k}{number of columns to print}
}
\description{
This function helps preventing Rstudio from printing out huge matrices
}
