% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_tcareg}
\alias{test_tcareg}
\title{Wrapper for Testing tcareg}
\usage{
test_tcareg(d, ...)
}
\arguments{
\item{d}{list of input matrixces, X, W, C1 and C2 required}

\item{...}{args passed to tcareg}
}
\value{
a \code{\link[=tcareg]{tcareg()}} fit
}
\description{
Wrapper for Testing tcareg
}
\keyword{internal}
