% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca_split.R
\name{split_input}
\alias{split_input}
\title{Split \code{\link[=tca]{tca()}} Input into Chunks by X}
\usage{
split_input(X, n_chunks = 1, shuffle = TRUE)
}
\arguments{
\item{X}{An \code{m} by \code{n} matrix of measurements of \code{m} features for \code{n} observations. Each column in \code{X} is assumed to be a mixture of \code{k} sources. Note that \code{X} must include row names and column names and that NA values are currently not supported. \code{X} should not include features that are constant across all observations.}

\item{n_chunks}{number of chunks to split X into. See Details.}

\item{shuffle}{randomly shuffle rows of X. This is used to make sure estimates
of tau from the data if requested is not confounded by the order of the features
if the data is run in chunks of X. Picking number of chunks equal to the
number of cores to parallel over is recommended.}
}
\value{
a split_input object that contains chunks of the X matrix by row.
}
\description{
Split the X matrix input into row chunks with equal number of features to
facilitate parallel processing over X for \code{\link[=tca]{tca()}}. Does not support
\code{refit_W = TRUE}.
}
\details{
Because of the assumption of independence between features of \code{\link[=tca]{tca()}}, splitting
the input of \code{\link[=tca]{tca()}} into chunks of features can potentially speed up the fit.
This is because we would only need to spin up the clusters once, run over all
chunks, and combine the result at the end.
}
\examples{
data <- test_data(30, 1000, 6, 1, 1, 0.01)
split_X <- split_input(X = data$X, n_chunks = 10)
split_X
dim(split_X)
}
